Bom_project_back/mob/requirements.txt
asgiref==3.2.10
cffi==1.14.0
cryptography==2.9.2
Django==3.0.7
django-filter==2.3.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
importlib-metadata==1.7.0
Markdown==3.2.2
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
zipp==3.1.0
